Refactor: user/edit API 경로변경
