+++ app/api/user/edit/route.ts
@@ -6,7 +6,6 @@ import { verifyAccessToken } from '@/infra/utils/jwt';
 import { SbUserRepo } from '@/infra/repositories/supabase/SbUserRepo';
 import { PutUserInfoUseCase } from '@/application/usecases/user/PutUserInfoUseCase';
 import { ValidationError } from '@/domain/error';
-import { MatchPasswordUsecase } from '@/application/usecases/user/MatchPasswordUsecase';
 
 export async function PUT(req: NextRequest) {
   try {
@@ -32,42 +31,6 @@ export async function PUT(req: NextRequest) {
         if(error instanceof ValidationError) {
             if(error.message === "nickName") return NextResponse.json({ status: 400, message: "nickName" })
             if(error.message === "phone_number") return NextResponse.json({ status: 400, message: "phone_number" })
-        }
-
-        if(error instanceof JsonWebTokenError) return NextResponse.json({ status: 401, message: "인증실패" });
-        
-        return NextResponse.json({status: 500, message: errorMessage});
-    }
-
-    return NextResponse.json({ status: 520, message: error });
-  }
-}
-
-
-export async function POST(req: NextRequest) {
-  try {
-    //request
-    const getHeaderToken = req.headers.get('authorization');
-    // Authorization 필드 확인영역
-    if (!getHeaderToken) throw new JsonWebTokenError("No token");
-    const access_token = getHeaderToken.replace('Bearer ', '');
-    const { user } = verifyAccessToken(access_token);
-
-    // body 필드 확인영역
-    const { password }: {password: string;} = await req.json();
-    const sbUserRepo = new SbUserRepo();
-    const matchPasswordUsecase = new MatchPasswordUsecase(sbUserRepo);
-    const data = await matchPasswordUsecase.excute(user.email, password);
-    console.log(data);
-
-    return NextResponse.json({ status: 200, data});
-  } catch (error) {
-    
-    if( error instanceof Error){
-        const errorMessage = error.message || "서버에러"
-
-        if(error instanceof ValidationError) {
-            if(error.message === "password") return NextResponse.json({ status: 400, message: "password" })
         }
 
         if(error instanceof JsonWebTokenError) return NextResponse.json({ status: 401, message: "인증실패" });
